Stop thread before deleting object reference
If setPeriodicCapture is called Visualizer will start a thread
and do periodic callbacks. When native_finalize is called the
reference to the class is deleted. If a callback happens after
this the reference is no longer valid and when used from the
callback method the vm crashes. With this patch the thread will
be terminated before deleting the reference.

Bug: 30684284

Depends-On: I3ea06243c44abd2f4e257199e1fb09386e94a132

Change-Id: I3ea06243c44abd2f4e257199e1fb09386e94a132
Signed-off-by: mydongistiny <jaysonedson@gmail.com>