Modified to perform the mute judgment in an valid index
There are minimum volume index values for audio streams.
In the volume setting of the stream, if the volume lower
than the minimum, it will fix the volume to minimum.
The function mutes the stream if the volume setting of
the stream is zero.

The problem is the mute judgment depends on the original volume.
It should depend on the valid volume.

Change-Id: I15f0909c491e4890230a1d7fb9ca3cb8394d9244
Signed-off-by: mydongistiny <jaysonedson@gmail.com>