ambient display: Fix volume key music control
Previously the dream service would only report whether or not it was dreaming.
Ambient display counts as dreaming.  When dreaming, the volume inputs are
passed back to the application which is not what we want for ambient
display.

Fix this by exposing isDozing() and allow volume keys to control music
when we are dreaming and have moved into doze mode (aka, sleeping).