Fix WFD source incoming persistent connection when device is P2P client
In the case when we receive an incoming connection from the
sink and we are client, we cannot retrieve all device
informations from getOwner() method which only returns
deviceAddress set in updatePersistentNetworks(), causing
an uncatched exception later in isSessionAvailable().

So in case of client, get devices informations
from mAvailableWifiDisplayPeers that contains all
the required data.

Change-Id: I73451176bf9391ab2088919c6bc35569eaa2f36a
Signed-off-by: Cedric Baudelet <cedric.baudelet@intel.com>
Signed-off-by: Zhiquan Liu <zhiquan.liu@intel.com>