SysUI: Wait for keyguard gone before starting activity
Some apps, I'm looking at you Chrome, don't like to do any work
if they are being started while the lock screen is shown, even
if it is being dismissed.  This patch adds another version of
startActivityDismissingKeyguard() which takes an extra boolean
parameter that allows to override the afterKeyguardGone behavior.
Normally that is only true if the resolver will be shown instead
of launching an app.  Overriding this allows us to get the exact
behavior we want without altering this behavior under normal
circumstances.

Change-Id: Ia804c7a5a73f2a4ad45593bcfcc192fedddf8d92
Signed-off-by: mydongistiny <jaysonedson@gmail.com>