Stop adjustConfigurationLw causing sleep/lock.
Currently PhoneWindowManager.adjustConfigurationLw
may trigger device sleep or screen lock when lid is closed.
adjustConfigurationLw is not expected to make a state change, since
the only caller is WindowManagerService.computeScreenConfigurationLocked,
whose role is only to compute configuration.
Closed lid event is handled in PhoneWindowManager.notifyLidSwitchChanged,
so we don't need the call to applyLidSwitchState here.

Change-Id: Ia39db96d3e19f71cbad571571a432011b9422a72
Signed-off-by: Takaki Makino <tmakino@google.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>