Added CloseGuard for Allocation
Bug: 27719830

This is only added for Allocation, which uses most significant amount of native
memory.

We could add this for other BaseObj too, for which we will have to add the
guard.open("destroy") call to the end of the constructors in a similar way.

To turn on warnings, apps have to add to their Activity.onCreate() method
the following code.

        StrictMode.setVmPolicy(new StrictMode.VmPolicy.Builder()
                               .detectLeakedClosableObjects()
                               .penaltyLog()
                               .build());

Change-Id: I77ed45239a60b85af5c811dee6c124fb53da9060
Signed-off-by: mydongistiny <jaysonedson@gmail.com>