Don't forget the connection of living keyguard service
Symptom:
Even though the correct password is input, keyguard cannot be unlocked.

Root cause:
This issue occurs when onServiceDisconnected event is delayed.
If onServiceDisconnected is called after onServiceConnected,
mKeyguardService is set to null.
KeyguardServiceDelegate loses the connection of keyguard service
and it cannot communicate with keyguard.
For fail safe, KeyguardServiceDelegate keeps showing KeyguardScrim
window which is unremovable keyguard screen.

Solution:
When binder is alive, it ignores onServiceDisconnected event
and keeps mKeyguardService.

Change-Id: Ic516407f6f6919c2c969038ad6e758a5d620cdf7
Signed-off-by: mydongistiny <jaysonedson@gmail.com>