Lockscreen shortcuts: don't allow specific activities
The picker allowed users to pick specific activities from an app that it
did not explicitly expose to the user.

This is not good UX as many activities are not meant to be user facing
and are expecting certain input. Users can pick apps themselves, or apps
can opt-in to shortcuts and publish things themselves.

Ticket: CYNGNOS-2684

Change-Id: I9199d442dbe9eaa4140a706166a6fde87bc8957c
Signed-off-by: Roman Birg <roman@cyngn.com>