WindowManagerService : Ensure we update both size/position
computeShownBounds is necessary to ensure shownFrame is up to
date when setSurfaceBoundariesLocked is invoked to ensure any needed
size and positional changes go as part of one transaction. Relying on
animateLocked to send changes results in size/position being sent as two
separate transactions and surface flinger drawing a frame which looks incorrect
according to what the window specified.

Change-Id: If881047c060c73db0037fda7da1f66a89dd45487
Signed-off-by: mydongistiny <jaysonedson@gmail.com>