Treat job deadline constraints as any other constraint
If a job has no constraints other than a deadline constraint, the
previous behavior would run the job as soon as the system could.  Now,
only run these jobs when the job expires.

The API side already treats the deadline as a constraint, and specifically
allows this:

  // Allow jobs with no constraints - What am I, a database?
  if (!mHasEarlyConstraint && !mHasLateConstraint && !mRequiresCharging &&
          !mRequiresDeviceIdle && mNetworkType == NETWORK_TYPE_NONE) {
          throw new IllegalArgumentException("You're trying to build a job with no " +
                  "constraints, this is not allowed.");
  }

This fixes the test
android.jobscheduler.cts.TimingConstraintsTest#testJobParameters_expiredDeadline.

Change-Id: Id7cb50abffab61cc1c4c4767d795f8c96eca4e32