Fix bug where foreground activity fails to cancel auth/enrollment
This fixes a bug where the currently authenticating application
can't cancel itself in onPause(). The problem is that some applications
can set FLAG_RESUME_WHILE_PAUSING (like the home screen), which changes
the life cycle of IMPORTANCE_FOREGROUND to be removed before onPause()
completes.  The result is the currently authenticating activity
can't cancel fingerprint authentication or enrollment.

The fix is to allow apps to always cancel themselves.

Fixes bug 30380234

Change-Id: If65c2028014da8d5129506d0b45cc6c1e25f7860
Signed-off-by: mydongistiny <jaysonedson@gmail.com>