Don't ensure configuration for activity we are resuming.
Regression introduced in ag/865988 when ASS.ensureActivitiesVisibleLocked()
was added to the end of AS.completePauseLocked() without specifing the
starting/resuming activity. If the starting/resuming activity isn't specified
the resuming activity can be relaunched with the current configuration which
might be different from the configuration the activity will end up in if it has
fixed orientation.
We now pass in the starting/resuming activity to avoid this.

Bug: 30509213
Change-Id: Idd0be84745bad83a1924e850b9b4020b927a78d8
Signed-off-by: mydongistiny <jaysonedson@gmail.com>