OMS7-N: Load missing style attributes from original style
Filtered out and optimized for use on the Overlay Manager Service
by: @nicholaschum

If a theme overrides a style in an app and does not include all
the attributes found in the original, it is possible for the app
to crash because it cannot locate an attribute it is expecting.

To overcome this we need to check the attributes in the original
style's bag and add any missing attributes to the current bag being
populated.  Currently we have only seen one theme that causes this
behavior, but there could be others out there or we may eventually
see this happen in the wild as apps update and the themes have not
had a chance to incorporate the new additions from the app.

Change-Id: Ib8579be08a9a6caaa84433713801de9fc97f753f
Signed-off-by: mydongistiny <jaysonedson@gmail.com>