Zygote: Workaround for detecting open file descriptors
Workaround for the zygote detecting open file descriptors when
the wrapper path is taken.

Add a permissive mode to the file descriptor table which will
ignore unknown or unexpected file descriptors. Pass down a new
zygote-internal debug flag for invoke-with and elide comparing
with the regular file descriptor table, using a temporary
permissive table instead (still using a table so that known
descriptors get reopened).

Bug: 32607405
Bug: 33668201
Test: m
Test: Device boots
Test: Special-purpose app test succeeds
Change-Id: Ib36d9488d141743cdba114875d945ee4e2000760
Signed-off-by: mydongistiny <jaysonedson@gmail.com>