Save "mState.restored" in onSaveInstanceState of FilesActivity
Symptom:
NPE occurred after rotating MUT in "Search result" screen.

Root cause:
As per createNewWindow method, FilesActivity will be started with
EXTRA_STACK intent extra when creating new window, then in onCreate
method, the refreshCurrentRootAndDirectory method will be called
since the "mState.stack" is not empty, it will refresh the content
using the state stack that is passed from EXTRA_STACK intent extra.

But, after FilesActivity is started, "mState.restored" is not assigned
as true even if state stack is ready now. It cause that after rotating
MUT in "Search result" screen, refreshCurrentRootAndDirectory method
is still called in onCreate method.

Then, the NPE occurred in cancelSearch method because mSearchView is
still null even if isSearching is true. mSearchView will be assigned
after onCreateOptionsMenu method is called.

Solution:
In onSaveInstanceState method, we assign "mState.restored" as true if
the "mState.stack" is not empty to ensure that after rotating MUT the
refreshCurrentRootAndDirectory method is not called in onCreate method.

It is same as normal case, do not new window and then rotate MUT in the
"Search result" screen.

Bug: 34593795

Change-Id: I8064b511d54bda9b4e4dcdff9719b3944865ef74
Signed-off-by: mydongistiny <jaysonedson@gmail.com>