Avoid the restart process for launching provider killed by AMS.
The case is that a content provider process killed by LowmemoryKiller while the
provider launching time. And AMS will restart it for the launching provider,and
reuse the ProcessRecord, it may be killed by AMS#updateOomAdjLocked in the end
of attachApplicationLocked for empty cached process.

After that the process will restart again, and may be killed angian, another
restart, another restart...

So, we should give time for process to launch its provider, avoid the restart
and kill circle.

Test: monkey test for 10h without related issue
Change-Id: If4be91faf804493f48d3d075799576e39561ea09
Signed-off-by: mydongistiny <jaysonedson@gmail.com>