DocumentsUI: morethe 40+ times crash in monkey test
RootCause: In RecyclerView source code,
The start method which is called by
smoothScrollToPosition method check whether
mTargetPosition value equal -1 or not, if equal
throw IllegalArgumentException, in stop method
mTargetPosition is assigned to -1. Under normal
circumestances, smoothScrollToPosition methed should
call setTargetPosition method, mTargetPosition is
assigned to normal value, but when the setTargetPosition
method is called after stop method, mTargetPosition is
assigned to -1. In monkey test, start the smoothScroll
and stop the smoothScroll faster could cause this issue.
The general operation of people will not be the case.

Solution: Use try/catch IllegalArgumentException to
avoid this case.

Change-Id: I9acd1edc861f0e2a8892591742d006295142875a
Signed-off-by: mydongistiny <jaysonedson@gmail.com>